Update waas-overview.md
The entire paragraph starting with "Basically, CBB is a configuration state ..." needs to be completely deleted or reworked as it is simply not correct. The Defer setting defines an intent, it does not define what is actually installed on a system and characterizing as such has lead to tons of confusion particularly when enabling this setting has all kinds of nasty side effects like enabling Windows Update dual-scan. The defer updates setting is for one things and one thing alone -- deferring feature updates from being installed from Windows Update for Business. That's it. Enabling this when you are not using WUfB **will** cause you issues.